% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{dataframe_handler}
\alias{dataframe_handler}
\title{dataframe result handler}
\usage{
dataframe_handler(res)
}
\arguments{
\item{res}{The character(1) vector returned by OmicIDXGraphqlClient$exec()}
}
\description{
A handler for OmicIDXGraphqlClient results that returns a
"maximally simplified" version of the data themselves, usually
as a possibly nested data frame.
}
