% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{data}
\name{OmicIDXQuery}
\alias{OmicIDXQuery}
\title{A container for OmicIDXR queries}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OmicIDXQuery
}
\description{
A container for OmicIDXR queries
}
\keyword{datasets}
