% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{raw_handler}
\alias{raw_handler}
\title{raw result handler}
\usage{
raw_handler(res)
}
\arguments{
\item{res}{The character(1) vector returned by OmicIDXGraphqlClient$exec()}
}
\description{
A handler for OmicIDXGraphqlClient results that simply returns
the JSON results from a query as an R list.
}
